hoist Filter to its own scope
