+++ frontend/src/components/__tests__/FilterDatePicker.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import DatePicker from '../DatePicker';
+import DatePicker from '../FilterDatePicker';
 
 // eslint-disable-next-line react/prop-types
 const RenderDatePicker = ({ onUpdateFilter = () => {}, id = 'id', query = null }) => (
