+++ frontend/src/components/__tests__/FilterItem.js
